// generated/src/app/api-models/stage-prem-pay-cash.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePremPayCash {
  batchId: string;
  transactionId: number;
  cashId: number;
  customerId: string;
  customerType: string;
  individualFirstName: string;
  individualMiddleName: string;
  individualLastName: string;
  indivIdCodeQualifier: string;
  indivIdentificationCode: string;
  checkEftDate: Date;
  checkEftReceiptDate: Date;
  orgIdCodeQualifier: string;
  orgIdCode: string;
  orgRefIdQualifier: string;
  orgRefIdCode: string;
  supplementalIdCodeQualifier: string;
  supplementalIdCode: string;
  companyCode: string;
  glRefCode: string;
  arTransactionAmount: number;
  arTransactionType: string;
  transactionStatus: string;
}